Use OR-Tools/MathOpt for quadratic problem resolution [ANT-2546] (#2574)
This PR adds the possibility for the user to choose quadratic (QP)
solvers other than sirius, dor adequacy problems.
It does so by interfacing Antares QP description
(PROBLEME_ANTARES_A_RESOUDRE) to the new OR-Tools MathOpt API.

The new "quadratic-solver" parameter allows the user to switch solver:
- setting the value "sirius" will make Antares behave like before, by
shunting OR-Tools and calling sirius directly
- setting another value will make Antares use OR-Tools via the MathOpt
API. For now, only
"[pdlp](https://developers.google.com/optimization/lp/pdlp_math)", and
SCIP are supported. However, SCIP seems to be broken (see
[here](https://github.com/google/or-tools/discussions/4497)). XPRESS
will be added once we can bump to OR-Tools v9.12 (currently having
issues in Windows)

**Breaking change:**
- "ortools-solver" parameter has been renamed to "linear-solver"
- "solver-parameters" parameter has been renamed to
"linear-solver-parameters"
- New parameters have been added : "quadratic-solver" and
"quadratic-solver-parameters" (note that the last parameter is a
placeholder, and that is not used for now)

---------

Signed-off-by: Peter Mitri <peter.mitri@rte-france.com>
Co-authored-by: Guillaume PIERRE <guillaume.pierre@rte-france.com>
Co-authored-by: Florian Omnès <florian.omnes@rte-france.com>