Factorise antlr4 lib linking resolution (#2929)
https://github.com/AntaresSimulatorTeam/Antares_Simulator/pull/2824/files#r2109244580

This pull request introduces a new `link_antlr4` macro to centralize the
logic for linking Antlr4 libraries based on the platform, and replaces
repetitive code with calls to this macro across multiple
`CMakeLists.txt` files. This change improves maintainability and reduces
duplication.

### Centralizing Antlr4 linking logic:

*
[`src/cmake/utils.cmake`](diffhunk://#diff-6b12360adc5f32cb85c6d0362b5aedb195c480a0feeda6eb08957d26c692bde6R62-R69):
Added the `link_antlr4` macro to encapsulate the logic for linking
Antlr4 libraries based on the platform (`MSVC` for Windows and static
linking for Linux).

### Refactoring `CMakeLists.txt` files:

*
[`src/expressions/antlr-interface/CMakeLists.txt`](diffhunk://#diff-2169f6d38efba395fd179669afa78b29d479cca384aad44ef5c2e6cc2993db57L19-R19):
Replaced the platform-specific Antlr4 linking logic with a call to the
new `link_antlr4` macro.
*
[`src/libs/antares/additionalConstraintRhsExpression/CMakeLists.txt`](diffhunk://#diff-e4d3d44a67d05cce3918477cf56c598bb6c1ae16c57b6fb2f9501ed157565b5fL19-R19):
Updated to use the `link_antlr4` macro for linking Antlr4 libraries.
*
[`src/libs/antares/scenarioGroupParser/scenarioBuilderExpression/CMakeLists.txt`](diffhunk://#diff-7d188cb9e54ac5efc041b8d8c469b5369f2b90059496c05b3c7bd5590a81fd11L20-R20):
Simplified Antlr4 library linking by utilizing the `link_antlr4` macro.