Fix logic for map parsing in yaml-cpp (#2992)
For model fields `connections` and `area-connections`, the right
behavior is
- Check that the node is a map
- Check that the map has the right size

Previously, if the node was a map, the check on the size was skipped
because of `&&`. With existing studies, we had this sequence
- `node.IsMap() = true`
- `!node.IsMap() = false`
- `!node.IsMap() && anything = false`
so the `size` condition is skipped whenever `node.IsMap() = true`. 

Also, `Antares::IO::Inputs::YmlSystem::AreaConnection` expects 3
elements, not 4.